Extract AnimationData to own file
