+++ Runtime/Scripts/Timeline/Internal.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7fff04c12a59e9141b703596b58fa5fc
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
