+++ Runtime/Scripts/Timeline/Internal/AnimationData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 569f31c28712cfa4da8a0d892bc0c6d0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
